refactor(knowledge): 重构嵌入器和搜索功能
-移除 Embedder 接口，直接使用 embedding.Embedder
- 更新 Search 接口，增加 sort 参数
- 重构 Embed 方法，提高代码复用性
-调整 Score 类型，统一为 float64
- 优化 MockVectorDB 和 MilvusVectorDB 的 Search 实现

+++ knowledge/types.go
@@ -49,7 +49,7 @@ type SearchOptions struct {
 	// 元数据过滤条件
 	Filters map[string]interface{} `json:"filters,omitempty"`
 	// 相似度阈值
-	Threshold float32 `json:"threshold,omitempty"`
+	Threshold float64 `json:"threshold,omitempty"`
 }
 
 // LoadOptions 加载选项配置
@@ -73,7 +73,7 @@ type SearchResult struct {
 	// 文档信息
 	Document Document `json:"document"`
 	// 相似度得分
-	Score float32 `json:"score"`
+	Score float64 `json:"score"`
 	// 匹配的分块（如果启用了分块）
 	Chunk *Chunk `json:"chunk,omitempty"`
 }
